Skip notebook with errors in ecosystem check
Summary
--

I've been noticing this failure in the formatter ecosystem check and decide to
look into it. We fail to parse the notebook because some of the `code` cells
have non-Python code in them. `ruff format` only reports one of these,
corresponding to a shell snippet, but `ruff check` emits some additional errors
about JS code later in the file too.

Test Plan
--

This PR