[WIP] Hide empty snippets for full-file diagnostics
Summary
--

This isn't really polished for review yet, but this is the other commit I wanted
to spin off from #19415.

This suppresses empty snippets for empty ranges at the very beginning of a
file, and for empty ranges in non-existent files. Ruff includes empty ranges for
IO errors, for example.

Test Plan
--

TODO add a `ruff_db` test with an example