avoid collecting `original_ranges`
We can just track this as the offset between the original source length and the
current or `result` length. This offset is given by:

```rust
let offset = result.text_len().to_usize() - index;
```

which when added to the original `update_ranges` call:

```rust
update_ranges(index + offset, tab_width);
```

simplifies to just `result.text_len()` (`index` is added and then subtracted).

The other slight nuance here is that in the original `update_ranges` call
locations, we would also need to subtract the length of the new character from
the `index` argument, so I instead opted just to move the calls to before we
added the new character. This makes `result.text_len()` alone exactly the value
we need.