Fix tests on 32-bit architectures
Summary
--

Fixes #19640. I'm not sure these are the exact fixes we really want, but I
reproduced the issue in a 32-bit Docker container and tracked down the causes,
so I figured I'd open a PR.

As I commented on the issue, the `goto_references` test depends on the iteration
order of the files in an `FxHashSet` in `Indexed`. In this case, we can just
sort the output in test code.

Similarly, the tuple case depended on the order of overloads inserted in an
`FxHashMap`. `FxOrderMap` seemed like a convenient drop-in replacement, but I
don't know if that will have other detrimental effects. I did have to change the
assertion for the tuple test, but I think it should now be stable across
architectures.

Test Plan
--

Running the tests in the aforementioned Docker container