PEP 639 license information
1. Declare licenses using keys `license` and `license-files`, as per PEP 639:
   * license:       SPDX license expression
   * license-files: list of license file glob patterns
2. Keep the Trove classifier in an attempt to remain compatible with tools
that do not support PEP 639 yet, such as pip-licenses. Hopefully the build
system will not complain about the redundant/obsolete classifier.

In theory, the build system maturin ≥ 1.9 supports PEP 639, which means
the proper licensing metadata should be included in packages:
	https://www.maturin.rs/changelog.html

In practice, maturin currently does not take license-files into account and
the LICENSE file is therefore missing from packages. I expect this maturin
bug to be fixed in uppcoming version 1.9.3.