fall back on diagnostic id for sarif
This one is complicated a bit by the `unique_rules` collection, which excludes
syntax errors. I added the new enum to avoid needing to change the definition of
`SarifRule::from` for `SecondaryCode`.

Initially it felt like a syntax error should also be a unique `SarifRule`, but
it doesn't have most of the other fields like `linter`, `summary`, or `url`, and
`name` and `code` would be the same.